// packages/core/src/types/id.ts
/**
 * Supported entity types for ID generation.
 */
export type EntityType =
  | 'account'
  | 'address'
  | 'apiKey'
  | 'authorizationCode'
  | 'contact'
  | 'currency'
  | 'customer'
  | 'department'
  | 'device'
  | 'email'
  | 'event'
  | 'feature'
  | 'featureFlag'
  | 'featureFlagOverride'
  | 'group'
  | 'importJob'
  | 'importJobRow'
  | 'importTemplate'
  | 'invoice'
  | 'log'
  | 'membership'
  | 'mobileNumber'
  | 'note'
  | 'notification'
  | 'oauthGrant'
  | 'orgFeature'
  | 'organization'
  | 'permission'
  | 'plan'
  | 'registeredApp'
  | 'record'
  | 'role'
  | 'session'
  | 'subscription'
  | 'team'
  | 'ticket'
  | 'user'
  | 'userFeature'
  | 'userProfile'